requirements.txt
Authlib==1.3.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pycparser==2.22
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==2.3.7
